Run Mocha on Webpack-generated bundle
